missing Debug impl'ns lint: `derive` suggestion only when available
Previously, the `missing_debug_implementations` lint's main message
instructed the user to consider deriving `Debug` or adding a manual
implementation. But isn't that kind of supplementary advice best relegated
to a note, rather than taking up space in the main message? (This is not to
mention that by using the diagnostic system's standard suggestion
machinery, we can make the suggestion available to tooling like the Rust
Language Server.) Furthermore, isn't giving such a context-insensitive
blanket statement a dereliction of our duty to our users to provide the
most specific, intelligent advice possible?—slapping `#[derive(Debug)]` on
a type whose fields aren't `Debug` _isn't going to work_, and we, the
compiler authors, are in a position to _know_ that, and refrain from
sending our loyal users on a fool's errand or a wild goose chase.