use fields in autotupling parse recovery
We introduce a new `PartialPResult`, which is like `PResult`, except
instead of the `Err` value being just a `DiagnosticBuilder`, it's a
tuple whose first element is of the `Ok` type and whose second element
is the diagnostic-builder: this provides a way to return to callers any
useful parsing work we did before hitting the error (at the cost of
having to sprinkle some awkward `.map_err`s around to convert between
`PResult` and `PartialPResult`).

`PartialPResult` is then used for more thorough recovery when a user
omits the parens from what we presume ought to have been a tuple
pattern: after emitting an error for the erroneous pattern, we can
continue working with the destructured fields as if the user had
correctly written a tuple, potentially catching more errors in
subsequent code without the user needing to recompile.