UI test: "elided" has just the one 'l'; exercise underscore lifetime
The files were named 'ellided-lifetimes', but the word "elided" only has
one 'l' (see, e.g., https://en.wiktionary.org/wiki/elide). We add a
couple lines to the test that exercise underscore lifetimes with the
elided-lifetimes-in-paths lint: in the current implementation, the lint
fires (as one can see in the expectation file), but it really shouldn't
(notice that the message says to use an underscore lifetime!); this is
to be fixed in another commit.