in which the elided-lifetimes-in-paths lint stops firing on `'_`
`is_elided` is true for both underscore (`'_`) and implicit (no
representation in source code) lifetimes, but we don't want to fire the
lint on the former, because the entire point of the lint is to suggest
changing the latter to the former (see the initial issue #45992).

It seems unfortunate for there to be ambiguity on whether the word
"elided" includes underscore lifetimes or not—the mandate of the
elided-lifetimes-in-paths lint seems to suggest it doesn't, whereas
the `is_elided` method seems to suggest it does—but it's beyond us to
resolve that in this commit. For now, let the message say "implicit"
for definiteness.

This relates to #52041.