structured suggestions for elided-lifetimes-in-paths lint
This was a doozy!

We only issue a suggestion when there's just one implicit lifetime,
because there can be situations where there are several (notice how
`elided_path_lifetimes` in librustc/hir/lowering.rs takes a number of
expected lifetimes as an argument), but we don't have an actionable
suggestion to offer (e.g., if there are two implicit lifetimes in a
return type, we hit E0106 "missing lifetime specifiers" as well as
(still) triggering this lint).

This has to do with #52041.