rename `elided_lifetimes_in_paths` lint to `hidden_lifetimes_in_types`
It's only with reluctance and sadness that we rename a lint that has
already been renamed once (#50879), but it seems worth it to pick the
best name now because since the lint is relatively new and has
heretofore been allow-by-default, the ecosystem breakage should be
minimal. (And—also sadly—the fact that the original implementation was
so buggy for so long testifies that not very many people are tuning up
the allow-by-default lints. Also, as always, lint capping prevents
lint changes from spreading contagiously to dependencies.)

The rationales here are that—

 • "hidden" is less potentially ambiguous than "elided", because this
   lint is specifically about angle-bracketed lifetime parameters,
   whereas the term "elided" has a strong precedent for also
   encompassing omitted lifetime names in reference ('&') types, which
   is not the concern of this lint, and

 • "types" is a more specific description of where the lint fires than
   "paths" (indeed, previous implementations of the lint used to fire
   on non-type paths in ways that proved to be erroneous
   false-positives, as evidenced by applications of the suggestion to
   use an anonymous lifetime (`'_`) resulting in code that didn't even
   parse)

This comes from discussion on #52069.