use `MIN`/`MAX` constant names in integer pattern coverage messages
While many programmers may intuitively appreciate the significance of
magic numbers like −2147483648, Rust is about empowering everyone to
build reliable and efficient software! It's a bit more legible to
print the constant names.

The `max_as_i128`, &c. methods strewn in libsyntax are a bit
unsightly, but I fear this is really the most natural way to solve the
problem.

Resolves #56393.